requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
streamlit>=1.30.0
pandas>=2.0.0
